Reorganize headers and create unit test for trimVelocityQueue
