+++ include/iarc7_motion/QuadVelocityController.hpp
@@ -13,7 +13,7 @@
 #include <geometry_msgs/Transform.h>
 #include <geometry_msgs/TransformStamped.h>
 #include <geometry_msgs/Vector3.h>
-#include "iarc7_motion/FeedForwardPid.hpp"
+#include "FeedForwardPid.hpp"
 #include "iarc7_msgs/Float64Stamped.h"
 #include <tf2_ros/transform_listener.h>
